backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
aiohttp==3.9.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# FusionRAG dependencies
langchain==0.0.350
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==1.1.0
numpy==1.24.3
pandas==2.0.3
tiktoken==0.5.1
rich==13.7.0

# Additional dependencies for async operations
asyncio-throttle==1.0.2
